fix compile warning

+++ mod_funasr.cpp
@@ -687,7 +687,7 @@ static void *init_fun_asr(switch_core_session_t *session, const switch_codec_imp
     pvt->asr_dec_vol = _asr_dec_vol ? switch_core_session_strdup(session, _asr_dec_vol) : nullptr;
     if (pvt->asr_dec_vol) {
         double db = strtod(pvt->asr_dec_vol, nullptr);
-        pvt->vol_multiplier = pow(10, db / 20);
+        pvt->vol_multiplier = (float)pow(10, db / 20);
     }
     switch_mutex_init(&pvt->mutex, SWITCH_MUTEX_NESTED, switch_core_session_get_pool(session));
